refactor(react-query-devtools): move QueryRow to subdirectory
